Reworked e2e test structure
