+++ .github/workflows/CI.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   # These tests run slowly, so run them in a separate job
-  e2e-tests:
+  e2e-tests-Parallel:
     strategy:
       matrix:
         node: [ 18 ]
@@ -72,8 +72,81 @@ jobs:
         if: steps.playwright-cache.outputs.cache-hit != 'true'
         run: npx playwright install chromium
 
-      - name: Run e2e tests
-        run: export ETH_RPC=e2e-test && yarn --cwd packages/commonwealth start & (yarn --cwd packages/commonwealth wait-server && yarn --cwd packages/commonwealth test-e2e)
+      - name: Run e2e tests (yarn test-e2e)
+        run: yarn --cwd packages/commonwealth e2e-start-server & (yarn --cwd packages/commonwealth wait-server && yarn --cwd packages/commonwealth test-e2e)
+        env:
+          PORT: 8080
+          USES_DOCKER_PGSQL: true
+
+      - name: Archive code coverage results
+        uses: actions/upload-artifact@v2
+        with:
+          name: code-coverage-report
+          path: coverage# These tests run slowly, so run them in a separate job
+
+  e2e-tests-serial:
+    strategy:
+      matrix:
+        node: [ 18 ]
+    name: Commonwealth E2E Test Serial
+    runs-on: ubuntu-latest
+
+    services:
+      postgres:
+        image: postgres
+        env:
+          POSTGRES_USER: commonwealth
+          POSTGRES_DB: commonwealth
+          POSTGRES_PASSWORD: edgeware
+        # Set health checks to wait until postgres has started
+        options: >-
+          --health-cmd pg_isready
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
+        ports:
+          # Maps tcp port 5432 on service container to the host
+          - 5432:5432
+      redis:
+        image: redis:latest
+        ports:
+          - 6379:6379
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Setup Node
+        uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node }}
+
+      - name: Cache dependencies
+        id: cache
+        uses: actions/cache@v3
+        with:
+          path: '**/node_modules'
+          key: ${{ matrix.node }}-${{ runner.os }}-modules-${{ hashFiles('**/yarn.lock') }}
+
+      - name: Install dependencies
+        run: yarn --ignore-engines
+
+      - name: Init Db
+        run: cd packages/commonwealth && npx sequelize db:migrate
+
+      - name: Cache browsers
+        id: playwright-cache
+        uses: actions/cache@v3
+        with:
+          path: '~/.cache/ms-playwright/chromium-1045'
+          key: hashFiles('~/.cache/ms-playwright/chromium-1045') }}
+
+      - name: Install playwright browser drivers
+        if: steps.playwright-cache.outputs.cache-hit != 'true'
+        run: npx playwright install chromium
+
+      - name: Run serial e2e tests (yarn test-e2e-serial)
+        run: yarn --cwd packages/commonwealth e2e-start-server & (yarn --cwd packages/commonwealth wait-server && yarn --cwd packages/commonwealth test-e2e-serial)
         env:
           PORT: 8080
           USES_DOCKER_PGSQL: true
