+++ packages/commonwealth/package.json
@@ -65,7 +65,8 @@
     "compress-images": "npx ts-node -T ./scripts/compressImages.ts",
     "gen-e2e": "npx playwright codegen",
     "e2e-start-server": "ETH_RPC=e2e-test yarn start",
-    "test-e2e": "TEST_ENV=playwright npx playwright test ./test/e2e/*",
+    "test-e2e-serial": "TEST_ENV=playwright npx playwright test --workers 1 ./test/e2e/e2eSerial/*",
+    "test-e2e": "TEST_ENV=playwright npx playwright test --workers 4 ./test/e2e/e2eRegular/*",
     "wait-server": "chmod +x ./scripts/wait-server.sh && ./scripts/wait-server.sh",
     "reset-frack-db": "heroku pg:copy commonwealth-beta::CW_READ_DB DATABASE_URL --app commonwealth-frack --confirm commonwealth-frack",
     "storybook": "storybook dev -p 6006",
