+++ packages/commonwealth/test/e2e/e2eRegular/landing.spec.ts
@@ -1,6 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { PORT } from '../../server/config';
-import { removeUser } from './utils/e2eUtils';
+import { PORT } from '../../../server/config';
 
 test.describe('Commonwealth Homepage', () => {
   test('Amount of bundles has not increased', async ({ page }) => {
@@ -131,35 +130,4 @@ test.describe('Commonwealth Homepage - Links', () => {
     await privacy.click();
     await page.waitForURL(`http://localhost:${PORT}/privacy`);
   });
-
-  test('Test Login', async ({ page }) => {
-    await page.goto('http://localhost:8080/');
-
-    await login(page);
-
-    await page.waitForSelector('.new-or-returning');
-    await page.getByText('New Account').click();
-    await page.getByText('Finish').click();
-    await page.waitForSelector('.username');
-    const element = await page.$('.username');
-    expect(element).toBeTruthy();
-  });
 });
-
-// Since we lazily import web3 in order to inject metamask into the window, it might not be available right away.
-// This allows us to wait until it becomes available by re-clicking the login button until it shows up.
-export async function login(page) {
-  await removeUser();
-
-  await page.getByText('Login').click();
-  await page.waitForSelector('.LoginDesktop');
-
-  let metaMaskIcon = await page.$$("text='Metamask'");
-  do {
-    await page.mouse.click(0, 0);
-    await page.getByText('Login').click();
-    metaMaskIcon = await page.$$("text='Metamask'");
-  } while (metaMaskIcon.length === 0);
-
-  await page.getByText('Metamask').click();
-}
